Move ParserHelper back into Core project
