+++ Source/AbsInt/AbsInt-NetCore.csproj
@@ -9,7 +9,6 @@
     <ProjectReference Include="..\BaseTypes\BaseTypes-NetCore.csproj" />
     <ProjectReference Include="..\CodeContractsExtender\CodeContractsExtender-NetCore.csproj" />
     <ProjectReference Include="..\Core\Core-NetCore.csproj" />
-    <ProjectReference Include="..\ParserHelper\ParserHelper-NetCore.csproj" />
   </ItemGroup>
 
 </Project>
