+++ Source/BoogieDriver/BoogieDriver.csproj
@@ -226,10 +226,6 @@
       <Project>{CF41E903-78EB-43BA-A355-E5FEB5ECECD4}</Project>
       <Name>Houdini</Name>
     </ProjectReference>
-    <ProjectReference Include="..\ParserHelper\ParserHelper.csproj">
-      <Project>{FCD3AC7F-9DFD-46C8-AB1E-09F0B0F16DC5}</Project>
-      <Name>ParserHelper</Name>
-    </ProjectReference>
     <ProjectReference Include="..\Predication\Predication.csproj">
       <Project>{AFAA5CE1-C41B-44F0-88F8-FD8A43826D44}</Project>
       <Name>Predication</Name>
