--- Source/ParserHelper/ParserHelper-NetCore.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <AssemblyName>BoogieParserHelper</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\CodeContractsExtender\CodeContractsExtender-NetCore.csproj" />
-  </ItemGroup>
-
-</Project>
