+++ Source/UnitTests/CoreTests/CoreTests-NetCore.csproj
@@ -13,7 +13,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\BaseTypes\BaseTypes-NetCore.csproj" />
     <ProjectReference Include="..\..\Core\Core-NetCore.csproj" />
-    <ProjectReference Include="..\..\ParserHelper\ParserHelper-NetCore.csproj" />
     <ProjectReference Include="..\TestUtil\TestUtil-NetCore.csproj" />
   </ItemGroup>
