+++ Source/UnitTests/TestUtil/TestUtil-NetCore.csproj
@@ -10,7 +10,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Core-NetCore.csproj" />
-    <ProjectReference Include="..\..\ParserHelper\ParserHelper-NetCore.csproj" />
   </ItemGroup>
 
 </Project>
